[Feature] Refactor GetTinyBlocksCount for better understanding.
