+++ contract/AElf.Contracts.Consensus.AEDPoS/AEDPoSContract_CacheFileds.cs
@@ -7,5 +7,6 @@ namespace AElf.Contracts.Consensus.AEDPoS
         private readonly Dictionary<long, Round> _rounds = new Dictionary<long, Round>();
         private readonly long _currentRoundNumber = 0;
         private string _processingBlockMinerPubkey;
+        private int _minimumBlocksCount = 0;
     }
 }
