+++ contract/AElf.Contracts.Consensus.AEDPoS/AEDPoSContract_HelpMethods_GetConsensusBehaviour.cs
@@ -37,7 +37,7 @@ namespace AElf.Contracts.Consensus.AEDPoS
             }
             else if (!isTimeSlotPassed)
             {
-                var blocksCount = GetTinyBlocksCount();
+                var blocksCount = GetMinimumBlocksCount();
                 if (minerInRound.ProducedTinyBlocks < blocksCount)
                 {
                     return AElfConsensusBehaviour.TinyBlock;
@@ -79,7 +79,7 @@ namespace AElf.Contracts.Consensus.AEDPoS
 
             if (currentRound.ExtraBlockProducerOfPreviousRound == publicKey && // If this miner is extra block producer of previous round,
                 Context.CurrentBlockTime < currentRound.GetStartTime() && // and currently the time is ahead of current round,
-                minerInRound.ProducedTinyBlocks < GetTinyBlocksCount() // make this miner produce some tiny blocks.
+                minerInRound.ProducedTinyBlocks < GetMinimumBlocksCount() // make this miner produce some tiny blocks.
             )
             {
                 return AElfConsensusBehaviour.TinyBlock;
