+++ contract/AElf.Contracts.Consensus.AEDPoS/AEDPoSContract_HelpMethods_GetConsensusCommand.cs
@@ -115,61 +115,6 @@ namespace AElf.Contracts.Consensus.AEDPoS
             }
         }
 
-        /// <summary>
-        /// Implemented GitHub PR #1952.
-        /// Adjust (mainly reduce) the count of tiny blocks produced by a miner each time to avoid too many forks.
-        /// </summary>
-        /// <returns></returns>
-        private int GetTinyBlocksCount()
-        {
-            TryToGetCurrentRoundInformation(out var currentRound, true);
-            var libRoundNumber = currentRound.ConfirmedIrreversibleBlockRoundNumber;
-            var libBlockHeight = currentRound.ConfirmedIrreversibleBlockHeight;
-            var currentHeight = Context.CurrentHeight;
-            var currentRoundNumber = currentRound.RoundNumber;
-
-            if (libRoundNumber == 0)
-            {
-                return AEDPoSContractConstants.MaximumTinyBlocksCount;
-            }
-
-            var (blockchainMiningStatus) =
-                new BlockchainMiningStatusEvaluator(
-                    libRoundNumber,
-                    libBlockHeight,
-                    currentRoundNumber,
-                    currentHeight);
-
-            Context.LogDebug(() => $"Current blockchain mining status: {blockchainMiningStatus.ToString()}");
-
-            // If R_LIB + 2 < R < R_LIB + 10 & H <= H_LIB + Y, CB goes to Min(L2/(R-R_LIB), CB0), while CT stays same as before.
-            if (blockchainMiningStatus == BlockchainMiningStatus.Abnormal)
-            {
-                var previousRoundMinedMinerList = State.MinedMinerListMap[currentRoundNumber.Sub(1)].Pubkeys;
-                var previousPreviousRoundMinedMinerList = State.MinedMinerListMap[currentRoundNumber.Sub(2)].Pubkeys;
-                var minersOfLastTwoRounds = previousRoundMinedMinerList
-                    .Intersect(previousPreviousRoundMinedMinerList).Count();
-                var count = Math.Min(AEDPoSContractConstants.MaximumTinyBlocksCount, minersOfLastTwoRounds
-                    .Div((int) currentRound.RoundNumber.Sub(libRoundNumber))
-                    .Add(1));
-                Context.LogDebug(() => $"Maximum blocks count tune to {count}");
-                return count;
-            }
-
-            //If R > R_LIB + 10 || H > H_LIB + Y, CB goes to 1, and CT goes to 0
-            if (blockchainMiningStatus == BlockchainMiningStatus.Severe)
-            {
-                // Fire an event to notify miner not package normal transaction.
-                Context.Fire(new IrreversibleBlockHeightUnacceptable
-                {
-                    DistanceToIrreversibleBlockHeight = currentHeight.Sub(libBlockHeight)
-                });
-                return 1;
-            }
-
-            return AEDPoSContractConstants.MaximumTinyBlocksCount;
-        }
-
         #region Get next block mining left milliseconds
 
         private void GetScheduleForUpdateValueWithoutPreviousInValue(Round currentRound,
@@ -304,7 +249,7 @@ namespace AElf.Contracts.Consensus.AEDPoS
             var producedTinyBlocksForPreviousRound =
                 minerInRound.ActualMiningTimes.Count(t => t < currentRoundStartTime);
 
-            var blocksCount = GetTinyBlocksCount();
+            var blocksCount = GetMinimumBlocksCount();
             if (minerInRound.ProducedTinyBlocks == blocksCount ||
                 minerInRound.ProducedTinyBlocks == blocksCount.Add(producedTinyBlocksForPreviousRound))
             {
